+++ src/main/java/com/projectj2ee/travel_server/repository/DestinationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DestinationRepository extends JpaRepository<Destination,Long> {
+public interface DestinationRepository extends JpaRepository<Destination,Integer> {
 }
