+++ src/main/java/com/projectj2ee/travel_server/repository/TourGuideRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TourGuideRepository extends JpaRepository<TourGuide,Long> {
+public interface TourGuideRepository extends JpaRepository<TourGuide,Integer> {
 }
